+++ tests/unit/projection-test.js
@@ -20,9 +20,11 @@ const BOOK_EXCERPT_PROJECTION_CLASS_PATH = 'com.example.bookstore.projection.Boo
 const NORM_BOOK_EXCERPT_PROJECTION_CLASS_PATH = 'com.example.bookstore.projection.book-excerpt';
 const BOOK_PREVIEW_PROJECTION_CLASS_PATH = 'com.example.bookstore.projection.BookPreview';
 const NORM_BOOK_PREVIEW_PROJECTION_CLASS_PATH = 'com.example.bookstore.projection.book-preview';
+const PROJECTED_AUTHOR_CLASS = 'com.example.bookstore.projectedType.ProjectedAuthor';
 const NORM_PROJECTED_AUTHOR_CLASS = 'com.example.bookstore.projected-type.projected-author';
-const PUBLISHER_CLASS = 'com.example.bookstore.publisher';
-const PROJECTED_PUBLISHER_CLASS = 'com.example.bookstore.projectedType.projectedPublisher';
+const PUBLISHER_CLASS = 'com.example.bookstore.Publisher';
+const NORM_PUBLISHER_CLASS = 'com.example.bookstore.publisher';
+const PROJECTED_PUBLISHER_CLASS = 'com.example.bookstore.projectedType.ProjectedPublisher';
 const NORM_PROJECTED_PUBLISHER_CLASS = 'com.example.bookstore.projected-type.projected-publisher';
 
 module('unit/projection', function(hooks) {
@@ -86,21 +88,21 @@ module('unit/projection', function(hooks) {
       models: {
         [NORM_BOOK_CLASS_PATH]: {},
         [NORM_BOOK_EXCERPT_PROJECTION_CLASS_PATH]: {
-          baseType: NORM_BOOK_CLASS_PATH,
+          baseType: BOOK_CLASS_PATH,
           attributes: ['title', 'author', 'year', 'publisher'],
         },
         [NORM_BOOK_PREVIEW_PROJECTION_CLASS_PATH]: {
-          baseType: NORM_BOOK_CLASS_PATH,
+          baseType: BOOK_CLASS_PATH,
           attributesTypes: {
-            publisher: NORM_PROJECTED_PUBLISHER_CLASS,
-            author: NORM_PROJECTED_AUTHOR_CLASS,
-            otherBooksInSeries: NORM_BOOK_PREVIEW_PROJECTION_CLASS_PATH,
+            publisher: PROJECTED_PUBLISHER_CLASS,
+            author: PROJECTED_AUTHOR_CLASS,
+            otherBooksInSeries: BOOK_PREVIEW_PROJECTION_CLASS_PATH,
           },
           // if you want to project an embedded model then it must have a type
           //  computedEmbeddedType
           attributes: ['title', 'author', 'chapter-1', 'year', 'publisher', 'otherBooksInSeries'],
         },
-        [PUBLISHER_CLASS]: {},
+        [NORM_PUBLISHER_CLASS]: {},
         // this schema must come with the parent schema
         [NORM_PROJECTED_AUTHOR_CLASS]: {
           attributes: ['location', 'name'],
@@ -1778,6 +1780,256 @@ module('unit/projection', function(hooks) {
     });
   });
 
+  module('Update projection property with resolved value', function(hooks) {
+    // properties for use for initial state
+    const BOOK_ID = 'isbn:9780439708181';
+    const PUBLISHER_ID = 'publisher-abc123';
+    const PUBLISHER_ID_NEW = 'publisher-abc123_new';
+    const PUBLISHER_URN = `urn:${PUBLISHER_CLASS}:${PUBLISHER_ID}`;
+
+    // intial old values
+    const PUBLISHER_NAME = 'MACMILLAN';
+    const PUBLISHER_LOCATION = 'Isle of Arran, Scotland';
+    const PUBLISHER_OWNER = 'Daniel and Alexander Macmillan';
+
+    // properties for use post-patch
+    const NEW_PUBLISHER_NAME = 'MACMILLAN NEW';
+    const NEW_PUBLISHER_LOCATION = 'London, England';
+    const NEW_PUBLISHER_OWNER = 'Holtzbrinck Publishing Group';
+    const NEW_PUBLISHER_URN = `urn:${PUBLISHER_CLASS}:${PUBLISHER_ID_NEW}`;
+
+    hooks.beforeEach(function() {
+      //Adding .setAttribute hook in schema
+      SchemaManager.schema.setAttribute = function(modelName, attr, value, schemaInterface) {
+        const baseModelName = this.computeBaseModelName(modelName);
+        if (
+          baseModelName &&
+          attr === 'publisher' &&
+          value &&
+          value.constructor &&
+          value.constructor.isModel
+        ) {
+          schemaInterface.setAttr(attr, NEW_PUBLISHER_URN);
+          return;
+        }
+
+        schemaInterface.setAttr(attr, value);
+      };
+
+      let { store } = this;
+
+      let baseRecord;
+      let projectedExcerpt;
+      let projectedPreview;
+
+      run(() => {
+        baseRecord = store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_CLASS_PATH,
+            attributes: {
+              publisher: PUBLISHER_URN,
+            },
+          },
+          included: [
+            {
+              id: PUBLISHER_ID,
+              type: PUBLISHER_CLASS,
+              attributes: {
+                name: PUBLISHER_NAME,
+                location: PUBLISHER_LOCATION,
+                owner: PUBLISHER_OWNER,
+              },
+            },
+            {
+              id: PUBLISHER_ID,
+              type: PROJECTED_PUBLISHER_CLASS,
+              attributes: {},
+            },
+          ],
+        });
+
+        projectedExcerpt = store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_EXCERPT_PROJECTION_CLASS_PATH,
+            attributes: {},
+          },
+        });
+
+        projectedPreview = store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_PREVIEW_PROJECTION_CLASS_PATH,
+            attributes: {},
+          },
+        });
+      });
+
+      this.records = {
+        baseRecord,
+        projectedExcerpt,
+        projectedPreview,
+      };
+    });
+
+    test('Updating property to another resolved value updates the base-record, other projections with new URN information using schema hook .setAttribute', function(assert) {
+      let { store } = this;
+      let { baseRecord, projectedExcerpt, projectedPreview } = this.records;
+
+      run(() => {
+        store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_EXCERPT_PROJECTION_CLASS_PATH,
+            attributes: {
+              publisher: PUBLISHER_URN,
+            },
+          },
+          included: [
+            {
+              id: PUBLISHER_ID,
+              type: PROJECTED_PUBLISHER_CLASS,
+              attributes: {
+                location: PUBLISHER_LOCATION,
+              },
+            },
+          ],
+        });
+
+        store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_PREVIEW_PROJECTION_CLASS_PATH,
+            attributes: {
+              publisher: PUBLISHER_URN,
+            },
+          },
+          included: [
+            {
+              id: PUBLISHER_ID,
+              type: PROJECTED_PUBLISHER_CLASS,
+              attributes: {
+                name: PUBLISHER_NAME,
+                loaction: PUBLISHER_LOCATION,
+              },
+            },
+          ],
+        });
+
+        // New publisher record
+        store.push({
+          data: {
+            id: PUBLISHER_ID_NEW,
+            type: PUBLISHER_CLASS,
+            attributes: {
+              name: NEW_PUBLISHER_NAME,
+              location: NEW_PUBLISHER_LOCATION,
+              owner: NEW_PUBLISHER_OWNER,
+            },
+          },
+        });
+      });
+
+      let newProjectedPublisherRecord = run(() =>
+        store.push({
+          data: {
+            id: PUBLISHER_ID_NEW,
+            type: PROJECTED_PUBLISHER_CLASS,
+            attributes: {
+              location: NEW_PUBLISHER_LOCATION,
+            },
+          },
+        })
+      );
+
+      // Value not changed in projection before setting new resolved value
+      assert.equal(
+        projectedPreview.get('publisher.id'),
+        PUBLISHER_ID,
+        'publisher.id is not updated'
+      );
+
+      assert.equal(
+        projectedPreview.get('publisher.location'),
+        PUBLISHER_LOCATION,
+        'publisher location is not updated'
+      );
+
+      assert.equal(
+        projectedPreview.get('publisher.name'),
+        PUBLISHER_NAME,
+        'publisher Name is not updated'
+      );
+
+      // Value not changed in base record before setting new resolved value
+      assert.equal(baseRecord.get('publisher.id'), PUBLISHER_ID, 'publisher.id is not updated');
+
+      assert.equal(
+        baseRecord.get('publisher.location'),
+        PUBLISHER_LOCATION,
+        'publisher location is not updated'
+      );
+
+      assert.equal(
+        baseRecord.get('publisher.owner'),
+        PUBLISHER_OWNER,
+        'publisher Owner is not updated'
+      );
+
+      assert.equal(
+        baseRecord.get('publisher.name'),
+        PUBLISHER_NAME,
+        'publisher Name is not updated'
+      );
+
+      // Set Resolved value
+      run(() => {
+        set(projectedExcerpt, 'publisher', newProjectedPublisherRecord);
+      });
+
+      // Value changed in projection after setting to new resolved value
+      assert.equal(
+        projectedPreview.get('publisher.id'),
+        PUBLISHER_ID_NEW,
+        'publisher.id is updated'
+      );
+
+      assert.equal(
+        projectedPreview.get('publisher.location'),
+        NEW_PUBLISHER_LOCATION,
+        'publisher location is updated'
+      );
+
+      assert.equal(
+        projectedPreview.get('publisher.name'),
+        NEW_PUBLISHER_NAME,
+        'publisher Name is updated'
+      );
+
+      // Value changed in base record after setting to new resolved value
+      assert.equal(baseRecord.get('publisher.id'), PUBLISHER_ID_NEW, 'publisher.id is updated');
+
+      assert.equal(
+        baseRecord.get('publisher.location'),
+        NEW_PUBLISHER_LOCATION,
+        'publisher location is updated'
+      );
+
+      assert.equal(
+        baseRecord.get('publisher.owner'),
+        NEW_PUBLISHER_OWNER,
+        'publisher Owner is updated'
+      );
+
+      assert.equal(
+        baseRecord.get('publisher.name'),
+        NEW_PUBLISHER_NAME,
+        'publisher Name is updated'
+      );
+    });
+  });
+
   skip(`Updates to a projection's non-whitelisted attributes do not cause a projection to be dirtied`, function() {});
 
   module('unloading/deleting records', function(hooks) {
@@ -2095,7 +2347,7 @@ module('unit/projection', function(hooks) {
       );
     });
 
-    skip('independently created projections of the same base-type and ID share their data', function(assert) {
+    test('independently created projections of the same base-type and ID share their data', function(assert) {
       let projectedPreview = run(() =>
         this.store.createRecord(BOOK_PREVIEW_PROJECTION_CLASS_PATH, {
           id: BOOK_ID,
@@ -2326,6 +2578,48 @@ module('unit/projection', function(hooks) {
         },
         /You cannot create a new projection for a pre-existing record/,
         '[TODO UPDATE THIS ASSERT] We throw the right assertion.'
+      );
+    });
+
+    test('.changedAttributes on a projection returns all changed properties', function(assert) {
+      let projectedExcerpt = run(() => {
+        return this.store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_EXCERPT_PROJECTION_CLASS_PATH,
+            attributes: {
+              title: BOOK_TITLE_1,
+              author: {
+                name: BOOK_AUTHOR_NAME_1,
+              },
+            },
+          },
+        });
+      });
+      let projectedPreview = run(() => {
+        return this.store.push({
+          data: {
+            id: BOOK_ID,
+            type: BOOK_PREVIEW_PROJECTION_CLASS_PATH,
+            attributes: {},
+          },
+        });
+      });
+
+      run(() => {
+        set(projectedPreview, 'title', BOOK_TITLE_2);
+        set(projectedPreview, 'author.name', BOOK_AUTHOR_NAME_2);
+      });
+
+      assert.deepEqual(
+        projectedExcerpt.changedAttributes(),
+        {
+          title: [BOOK_TITLE_1, BOOK_TITLE_2],
+          author: {
+            name: [BOOK_AUTHOR_NAME_1, BOOK_AUTHOR_NAME_2],
+          },
+        },
+        'Expected changed attributes to be correctly returned'
       );
     });